// src/ast/types.ts
export declare interface ArrayExpression extends Node_2 {
  type: NodeTypes.JS_ARRAY_EXPRESSION
  elements: Array<string | JSChildNode>
}

export declare function assert(condition: boolean, msg?: string): void

export declare interface AssignmentExpression extends Node_2 {
  type: NodeTypes.JS_ASSIGNMENT_EXPRESSION
  left: SimpleExpressionNode
  right: JSChildNode
}

export declare interface AttributeNode extends Node_2 {
  type: NodeTypes.ATTRIBUTE
  name: string
  value: TextNode | undefined
}
// ...
export declare interface BlockStatement extends Node_2 {
  type: NodeTypes.JS_BLOCK_STATEMENT
  body: (JSChildNode | IfStatement)[]
}
// ...
export declare interface CacheExpression extends Node_2 {
  type: NodeTypes.JS_CACHE_EXPRESSION
  index: number
  value: JSChildNode
  isVNode: boolean
}

export declare interface CallExpression extends Node_2 {
  type: NodeTypes.JS_CALL_EXPRESSION
  callee: string | symbol
  arguments: (string | symbol | JSChildNode | SSRCodegenNode | TemplateChildNode | TemplateChildNode[])[]
}
// ...
export declare interface CommentNode extends Node_2 {
  type: NodeTypes.COMMENT
  content: string
}

export declare interface CompilerError extends SyntaxError {
  code: number
  loc?: SourceLocation
}

export declare type CompilerOptions = ParserOptions & TransformOptions & CodegenOptions

export declare interface ComponentNode extends BaseElementNode {
  tagType: ElementTypes.COMPONENT
  codegenNode: VNodeCall | CacheExpression | undefined
  ssrCodegenNode?: CallExpression
}

export declare interface CompoundExpressionNode extends Node_2 {
  type: NodeTypes.COMPOUND_EXPRESSION
  children: (SimpleExpressionNode | CompoundExpressionNode | InterpolationNode | TextNode | string | symbol)[]
  /**
     * an expression parsed as the params of a function will track
     * the identifiers declared inside the function body.
     */
  identifiers?: string[]
}

export declare interface ConditionalDynamicSlotNode extends ConditionalExpression {
  consequent: DynamicSlotNode
  alternate: DynamicSlotNode | SimpleExpressionNode
}

export declare interface ConditionalExpression extends Node_2 {
  type: NodeTypes.JS_CONDITIONAL_EXPRESSION
  test: JSChildNode
  consequent: JSChildNode
  alternate: JSChildNode
  newline: boolean
}

/**
 * Static types have several levels.
 * Higher levels implies lower levels. e.g. a node that can be stringified
 * can always be hoisted and skipped for patch.
 */
export declare const enum ConstantTypes {
  NOT_CONSTANT = 0,
  CAN_SKIP_PATCH = 1,
  CAN_HOIST = 2,
  CAN_STRINGIFY = 3
}
// ...
export declare interface DirectiveNode extends Node_2 {
  type: NodeTypes.DIRECTIVE
  name: string
  exp: ExpressionNode | undefined
  arg: ExpressionNode | undefined
  modifiers: string[]
  /**
     * optional property to cache the expression parse result for v-for
     */
  parseResult?: ForParseResult
}
// ...
export declare type ElementNode = PlainElementNode | ComponentNode | SlotOutletNode | TemplateNode
// ...
export declare type ExpressionNode = SimpleExpressionNode | CompoundExpressionNode

export declare function findDir(node: ElementNode, name: string | RegExp, allowEmpty?: boolean): DirectiveNode | undefined

export declare function findProp(node: ElementNode, name: string, dynamicOnly?: boolean, allowEmpty?: boolean): ElementNode['props'][0] | undefined

export declare interface ForCodegenNode extends VNodeCall {
  isBlock: true
  tag: typeof FRAGMENT
  props: undefined
  children: ForRenderListExpression
  patchFlag: string
  disableTracking: boolean
}

export declare interface ForIteratorExpression extends FunctionExpression {
  returns: BlockCodegenNode
}

export declare interface ForNode extends Node_2 {
  type: NodeTypes.FOR
  source: ExpressionNode
  valueAlias: ExpressionNode | undefined
  keyAlias: ExpressionNode | undefined
  objectIndexAlias: ExpressionNode | undefined
  parseResult: ForParseResult
  children: TemplateChildNode[]
  codegenNode?: ForCodegenNode
}

declare interface ForParseResult {
  source: ExpressionNode
  value: ExpressionNode | undefined
  key: ExpressionNode | undefined
  index: ExpressionNode | undefined
}

export declare interface ForRenderListExpression extends CallExpression {
  callee: typeof RENDER_LIST
  arguments: [ExpressionNode, ForIteratorExpression]
}

export declare const FRAGMENT: unique symbol

export declare interface FunctionExpression extends Node_2 {
  type: NodeTypes.JS_FUNCTION_EXPRESSION
  params: ExpressionNode | string | (ExpressionNode | string)[] | undefined
  returns?: TemplateChildNode | TemplateChildNode[] | JSChildNode
  body?: BlockStatement | IfStatement
  newline: boolean
  /**
     * This flag is for codegen to determine whether it needs to generate the
     * withScopeId() wrapper
     */
  isSlot: boolean
}
// ...
export declare interface IfBranchNode extends Node_2 {
  type: NodeTypes.IF_BRANCH
  condition: ExpressionNode | undefined
  children: TemplateChildNode[]
  userKey?: AttributeNode | DirectiveNode
}

export declare interface IfConditionalExpression extends ConditionalExpression {
  consequent: BlockCodegenNode
  alternate: BlockCodegenNode | IfConditionalExpression
}

export declare interface IfNode extends Node_2 {
  type: NodeTypes.IF
  branches: IfBranchNode[]
  codegenNode?: IfConditionalExpression | CacheExpression
}

export declare interface IfStatement extends Node_2 {
  type: NodeTypes.JS_IF_STATEMENT
  test: ExpressionNode
  consequent: BlockStatement
  alternate: IfStatement | BlockStatement | ReturnStatement | undefined
}

declare interface ImportItem {
  exp: string | ExpressionNode
  path: string
}

declare type InferCodegenNodeType<T> = T extends typeof RENDER_SLOT ? RenderSlotCall : CallExpression

export declare function injectProp(node: VNodeCall | RenderSlotCall, prop: Property, context: TransformContext): void

export declare interface InterpolationNode extends Node_2 {
  type: NodeTypes.INTERPOLATION
  content: ExpressionNode
}
// ...
export declare type JSChildNode = VNodeCall | CallExpression | ObjectExpression | ArrayExpression | ExpressionNode | FunctionExpression | ConditionalExpression | CacheExpression | AssignmentExpression | SequenceExpression
// ...
declare interface Node_2 {
  type: NodeTypes
  loc: SourceLocation
}
export { Node_2 as Node }

export declare type NodeTransform = (node: RootNode | TemplateChildNode, context: TransformContext) => void | (() => void) | (() => void)[]

export enum NodeTypes {
  ROOT = 0,
  ELEMENT = 1,
  TEXT = 2,
  COMMENT = 3,
  SIMPLE_EXPRESSION = 4,
  INTERPOLATION = 5,
  ATTRIBUTE = 6,
  DIRECTIVE = 7,
  COMPOUND_EXPRESSION = 8,
  IF = 9,
  IF_BRANCH = 10,
  FOR = 11,
  TEXT_CALL = 12,
  VNODE_CALL = 13,
  JS_CALL_EXPRESSION = 14,
  JS_OBJECT_EXPRESSION = 15,
  JS_PROPERTY = 16,
  JS_ARRAY_EXPRESSION = 17,
  JS_FUNCTION_EXPRESSION = 18,
  JS_CONDITIONAL_EXPRESSION = 19,
  JS_CACHE_EXPRESSION = 20,
  JS_BLOCK_STATEMENT = 21,
  JS_TEMPLATE_LITERAL = 22,
  JS_IF_STATEMENT = 23,
  JS_ASSIGNMENT_EXPRESSION = 24,
  JS_SEQUENCE_EXPRESSION = 25,
  JS_RETURN_STATEMENT = 26
}

export declare const noopDirectiveTransform: DirectiveTransform

export declare interface ObjectExpression extends Node_2 {
  type: NodeTypes.JS_OBJECT_EXPRESSION
  properties: Array<Property>
}
// ...
export declare interface PlainElementNode extends BaseElementNode {
  tagType: ElementTypes.ELEMENT
  codegenNode: VNodeCall | SimpleExpressionNode | CacheExpression | undefined
  ssrCodegenNode?: TemplateLiteral
}
// ...
export declare interface RootNode extends Node_2 {
  type: NodeTypes.ROOT
  children: TemplateChildNode[]
  helpers: symbol[]
  components: string[]
  directives: string[]
  hoists: (JSChildNode | null)[]
  imports: ImportItem[]
  cached: number
  temps: number
  ssrHelpers?: symbol[]
  codegenNode?: TemplateChildNode | JSChildNode | BlockStatement | undefined
}

export declare interface SequenceExpression extends Node_2 {
  type: NodeTypes.JS_SEQUENCE_EXPRESSION
  expressions: JSChildNode[]
}
// ...
export declare interface SimpleExpressionNode extends Node_2 {
  type: NodeTypes.SIMPLE_EXPRESSION
  content: string
  isStatic: boolean
  constType: ConstantTypes
  /**
     * Indicates this is an identifier for a hoist vnode call and points to the
     * hoisted node.
     */
  hoisted?: JSChildNode
  /**
     * an expression parsed as the params of a function will track
     * the identifiers declared inside the function body.
     */
  identifiers?: string[]
}

export declare type SlotFnBuilder = (slotProps: ExpressionNode | undefined, slotChildren: TemplateChildNode[], loc: SourceLocation) => FunctionExpression

export declare interface SlotFunctionExpression extends FunctionExpression {
  returns: TemplateChildNode[]
}

export declare interface SlotOutletNode extends BaseElementNode {
  tagType: ElementTypes.SLOT
  codegenNode: RenderSlotCall | CacheExpression | undefined
  ssrCodegenNode?: CallExpression
}
// ...
export declare interface SourceLocation {
  start: Position
  end: Position
  source: string
}

export declare type SSRCodegenNode = BlockStatement | TemplateLiteral | IfStatement | AssignmentExpression | ReturnStatement | SequenceExpression

export declare type StructuralDirectiveTransform = (node: ElementNode, dir: DirectiveNode, context: TransformContext) => void | (() => void)

export declare const SUSPENSE: unique symbol

export declare const TELEPORT: unique symbol

export declare type TemplateChildNode = ElementNode | InterpolationNode | CompoundExpressionNode | TextNode | CommentNode | IfNode | IfBranchNode | ForNode | TextCallNode

export declare interface TemplateLiteral extends Node_2 {
  type: NodeTypes.JS_TEMPLATE_LITERAL
  elements: (string | JSChildNode)[]
}

export declare interface TemplateNode extends BaseElementNode {
  tagType: ElementTypes.TEMPLATE
  codegenNode: undefined
}

export declare type TemplateTextChildNode = TextNode | InterpolationNode | CompoundExpressionNode

export declare interface TextCallNode extends Node_2 {
  type: NodeTypes.TEXT_CALL
  content: TextNode | InterpolationNode | CompoundExpressionNode
  codegenNode: CallExpression | SimpleExpressionNode
}

export declare const enum TextModes {
  DATA = 0,
  RCDATA = 1,
  RAWTEXT = 2,
  CDATA = 3,
  ATTRIBUTE_VALUE = 4
}

export declare interface TextNode extends Node_2 {
  type: NodeTypes.TEXT
  content: string
}
// ...
export declare interface VNodeCall extends Node_2 {
  type: NodeTypes.VNODE_CALL
  tag: string | symbol | CallExpression
  props: PropsExpression | undefined
  children: TemplateChildNode[] | TemplateTextChildNode | SlotsExpression | ForRenderListExpression | undefined
  patchFlag: string | undefined
  dynamicProps: string | undefined
  directives: DirectiveArguments | undefined
  isBlock: boolean
  disableTracking: boolean
}